requirements.txt
pyyaml>=6.0.1
psutil>=5.9.8
PyQt5>=5.15.11
qtawesome>=1.3.1
pyttsx3>=2.90
opencv-python>=4.10.0
pyaudio>=0.2.14
numpy>=1.26.4
matplotlib>=3.9.2
faster-whisper>=0.10.0
opencc>=1.1.7
PyMySQL>=1.1.1
pyserial>=3.5
websockets>=12.0
